move to niimg.py
