+++ nilearn/_utils/niimg.py
@@ -14,6 +14,21 @@ import nibabel
 from .compat import _basestring
 
 
+def get_data(img):
+    # copy-pasted from https://github.com/nipy/nibabel/blob/de44a105c1267b07ef9e28f6c35b31f851d5a005/nibabel/dataobj_images.py#L204
+    # get_data is removed from nibabel because:
+    # see https://github.com/nipy/nibabel/wiki/BIAP8
+    caching = "fill"
+    if caching not in ('fill', 'unchanged'):
+        raise ValueError('caching value should be "fill" or "unchanged"')
+    if img._data_cache is not None:
+        return img._data_cache
+    data = np.asanyarray(img._dataobj)
+    if caching == 'fill':
+        img._data_cache = data
+    return data
+
+
 def _safe_get_data(img, ensure_finite=False):
     """ Get the data in the image without having a side effect on the
         Nifti1Image object
@@ -40,7 +55,7 @@ def _safe_get_data(img, ensure_finite=False):
     # that's why we invoke a forced call to the garbage collector
     gc.collect()
 
-    data = img.get_data()
+    data = get_data(img)
     if ensure_finite:
         non_finite_mask = np.logical_not(np.isfinite(data))
         if non_finite_mask.sum() > 0: # any non_finite_mask values?
@@ -113,16 +128,16 @@ def load_niimg(niimg, dtype=None):
                         " not compatible with nibabel format:\n"
                         + short_repr(niimg))
 
-    dtype = _get_target_dtype(niimg.get_data().dtype, dtype)
+    dtype = _get_target_dtype(get_data(niimg).dtype, dtype)
 
     if dtype is not None:
         # Copyheader and set dtype in header if header exists
         if niimg.header is not None:
-            niimg = new_img_like(niimg, niimg.get_data().astype(dtype),
+            niimg = new_img_like(niimg, get_data(niimg).astype(dtype),
                                 niimg.affine, copy_header=True)
             niimg.header.set_data_dtype(dtype)        
         else:
-            niimg = new_img_like(niimg, niimg.get_data().astype(dtype),
+            niimg = new_img_like(niimg, get_data(niimg).astype(dtype),
                                 niimg.affine)
 
     return niimg
