+++ nilearn/image/image.py
@@ -20,24 +20,9 @@ from .. import signal
 from .._utils import (check_niimg_4d, check_niimg_3d, check_niimg, as_ndarray,
                       _repr_niimgs)
 from .._utils.niimg_conversions import _index_img, _check_same_fov
-from .._utils.niimg import _safe_get_data
+from .._utils.niimg import _safe_get_data, get_data
 from .._utils.compat import _basestring
 from .._utils.param_validation import check_threshold
-
-
-def get_data(img):
-    # copy-pasted from https://github.com/nipy/nibabel/blob/de44a105c1267b07ef9e28f6c35b31f851d5a005/nibabel/dataobj_images.py#L204
-    # get_data is removed from nibabel because:
-    # see https://github.com/nipy/nibabel/wiki/BIAP8
-    caching = "fill"
-    if caching not in ('fill', 'unchanged'):
-        raise ValueError('caching value should be "fill" or "unchanged"')
-    if img._data_cache is not None:
-        return img._data_cache
-    data = np.asanyarray(img._dataobj)
-    if caching == 'fill':
-        img._data_cache = data
-    return data
 
 
 def high_variance_confounds(imgs, n_confounds=5, percentile=2.,
@@ -96,7 +81,7 @@ def high_variance_confounds(imgs, n_confounds=5, percentile=2.,
     else:
         # Load the data only if it doesn't need to be masked
         imgs = check_niimg_4d(imgs)
-        sigs = as_ndarray(imgs.get_data())
+        sigs = as_ndarray(get_data(imgs))
         # Not using apply_mask here saves memory in most cases.
         del imgs  # help reduce memory consumption
         sigs = np.reshape(sigs, (-1, sigs.shape[-1])).T
@@ -281,7 +266,7 @@ def smooth_img(imgs, fwhm):
     for img in imgs:
         img = check_niimg(img)
         affine = img.affine
-        filtered = _smooth_array(img.get_data(), affine, fwhm=fwhm,
+        filtered = _smooth_array(get_data(img), affine, fwhm=fwhm,
                                  ensure_finite=True, copy=True)
         ret.append(new_img_like(img, filtered, affine, copy_header=True))
 
@@ -323,7 +308,7 @@ def _crop_img_to(img, slices, copy=True):
 
     img = check_niimg(img)
 
-    data = img.get_data()
+    data = get_data(img)
     affine = img.affine
 
     cropped_data = data[tuple(slices)]
@@ -383,7 +368,7 @@ def crop_img(img, rtol=1e-8, copy=True, pad=True, return_offset=False):
     """
 
     img = check_niimg(img)
-    data = img.get_data()
+    data = get_data(img)
     infinity_norm = max(-data.min(), data.max())
     passes_threshold = np.logical_or(data < -rtol * infinity_norm,
                                      data > rtol * infinity_norm)
@@ -463,7 +448,7 @@ def _compute_mean(imgs, target_affine=None,
         target_affine=target_affine, target_shape=target_shape,
         copy=False)
     affine = mean_data.affine
-    mean_data = mean_data.get_data()
+    mean_data = get_data(mean_data)
 
     if smooth:
         nan_mask = np.isnan(mean_data)
@@ -580,7 +565,7 @@ def swap_img_hemispheres(img):
     img = reorder_img(img)
 
     # create swapped nifti object
-    out_img = new_img_like(img, img.get_data()[::-1], img.affine,
+    out_img = new_img_like(img, get_data(img)[::-1], img.affine,
                            copy_header=True)
 
     return out_img
@@ -987,7 +972,7 @@ def clean_img(imgs, sessions=None, detrend=True, standardize=True,
     if mask_img is not None:
         signals = masking.apply_mask(imgs_, mask_img)
     else:
-        signals = imgs_.get_data().reshape(-1, imgs_.shape[-1]).T
+        signals = get_data(imgs_).reshape(-1, imgs_.shape[-1]).T
 
     # Clean signal
     data = signal.clean(
